Module Monadlib

module Applicative : sig ... end

Applicative functors are data structures over one can map a function.

module Collection : sig ... end

module Continuation : sig ... end

module Identity : sig ... end

A trivial monad which brings into view some useful functions from Monad.S for simple types 'a.

module Lazy : sig ... end

module LazyList : sig ... end

An alternative to the batteries lazy list library, with generally improved laziness.

module LazyPlus : sig ... end

module List : sig ... end

module Monad : sig ... end

The monad library.

module MonadPlus : sig ... end

module Monoid : sig ... end

module Option : sig ... end

module Reader : sig ... end

module Result : sig ... end

module Retry : sig ... end

module State : sig ... end

module Stream : sig ... end

module TagTree : sig ... end

Tagged logic trees which are simplified when combined. When the contents are interpreted as a bag conditional values, should be a monad.

module Writer : sig ... end