Module Monadlib.Collection

module type T = sig ... end

Monads for collections. Stream is the current use-case for this, since it is parameterised on a collection monad (like list).

Transformers on Collections

Sometimes, you might want to transform a collection monad, in such a way that functions such as BaseCollectionM.unique behave in a sensible way by regarding None as smaller than any Some. Each transformer provides a function cmp_on with which the collection functions such as BaseCollectionM.difference are implemented. We also provide the list function for transformers.

module MakeOpt (C : T) : sig ... end