Module Monadlib.LazyPlus

module type T = sig ... end

S is another base module useful when the monad is a lazy data structure. We then allow the plus operation to be non-strict in its second argument, which makes it possible to use functions such as Monad.lsum lazily. This is what you want for lazy lists.

module type S = sig ... end

This is the counterpart for the lazy version of BasePlus.

module Make (M : T) : S with type 'a m = 'a M.m

module LazyList : S with type 'a m = 'a LazyList.t