Module Ap.Infix

val (<@>) : ( 'a -> 'b ) -> 'a m -> 'b m

Alias for map. f <@> x applies a pure function f to a monadic value x.

val (<*>) : ( 'a -> 'b ) m -> 'a m -> 'b m

f <*> x applies a monadic function f to a monadic value x. This operator is usually used in conjunction with <@> in order to accumulate monadic parameters. Example: max <@> return 1 <*> return 2 will yield return 2. Here max has arity 2, so we use <*> to add a second monadic parameter.

val (<*) : 'a m -> 'b m -> 'a m

x <* y ignores the result of y.

val (*>) : 'a m -> 'b m -> 'b m

x <* y ignores the result of x.