Module MonadPlus.Make

Parameters

module M : T

Signature

include T with type 'a m = 'a M.m

include BatInterfaces.Monad with type 'a m = 'a M.m

type 'a m = 'a M.m

val zero : unit -> 'a m

val plus : 'a m -> 'a m -> 'a m

val null : 'a m -> bool

null x implies that x is zero. If you do not want to or cannot answer whether a given x is zero, then null x should be false. I have provided this so that streams can be implemented more efficiently.

include Monad.S with type 'a m := 'a m

include BatInterfaces.Monad with type 'a m := 'a m

val bind : 'a m -> ('a -> 'b m) -> 'b m

module Ap : Applicative.S with type 'a m := 'a m

include Applicative.S with type 'a m := 'a m

include Applicative.T with type 'a m := 'a m

val return : 'a -> 'a m

val (<*>) : ('a -> 'b) m -> 'a m -> 'b m

val map : ('a -> 'b) -> 'a m -> 'b m

val fmap : ('a -> 'b) m -> 'a -> 'b m

val iter : ('a -> unit) -> 'a m -> unit

val (<@>) : ('a -> 'b) -> 'a m -> 'b m

Alias for map

val (<*>) : ('a -> 'b) m -> 'a m -> 'b m

val (<*) : 'a m -> 'b m -> 'a m

val (*>) : 'a m -> 'b m -> 'b m

val ignore : 'a m -> unit m

val map2 : ('a -> 'b -> 'c) -> 'a m -> 'b m -> 'c m

val map3 : ('a -> 'b -> 'c -> 'd) -> 'a m -> 'b m -> 'c m -> 'd m

val map4 : ('a -> 'b -> 'c -> 'd -> 'e) -> 'a m -> 'b m -> 'c m -> 'd m -> 'e m

List functions

val sequence : (unit -> 'a m) list -> 'a list m

val sequence_unit : (unit -> unit m) list -> unit m

val list_map : ('a -> 'b m) -> 'a list -> 'b list m

val list_iter : ('a -> unit m) -> 'a list -> unit m

val list_filter : ('a -> bool m) -> 'a list -> 'a list m

val list_filter_map : ('a -> 'b option m) -> 'a list -> 'b list m

Option functions

val option_map : ('a -> 'b m) -> 'a option -> 'b option m

Boolean function

val optional : bool -> (unit -> 'a m) -> 'a option m

val (>>=) : 'a m -> ('a -> 'b m) -> 'b m

val (>=>) : ('a -> 'b m) -> ('b -> 'c m) -> 'a -> 'c m

val (<=<) : ('b -> 'c m) -> ('a -> 'b m) -> 'a -> 'c m

include module type of Ap.Infix

val (<@>) : ('a -> 'b) -> 'a m -> 'b m

Alias for map

val (<*>) : ('a -> 'b) m -> 'a m -> 'b m

val (<*) : 'a m -> 'b m -> 'a m

val (*>) : 'a m -> 'b m -> 'b m

module Bindings : sig ... end

include module type of Bindings

val let* : 'a m -> ('a -> 'b m) -> 'b m

include module type of Ap.Bindings

val let+ : 'a m -> ('a -> 'b) -> 'b m

val and+ : 'a m -> 'b m -> ('a * 'b) m

val join : 'a m m -> 'a m

Note that Ap.list_map is overwritten with list_map_s

val list_fold_left : ('a -> 'b -> 'a m) -> 'a -> 'b list -> 'a m

val list_fold_right : ('a -> 'b -> 'b m) -> 'b -> 'a list -> 'b m

module Infix : sig ... end

include module type of Infix

val (++) : 'a m -> 'a m -> 'a m

val (+?) : 'a m option -> 'a m -> 'a m

val (>>=) : 'a m -> ('a -> 'b m) -> 'b m

val (>=>) : ('a -> 'b m) -> ('b -> 'c m) -> 'a -> 'c m

val (<=<) : ('b -> 'c m) -> ('a -> 'b m) -> 'a -> 'c m

val (<@>) : ('a -> 'b) -> 'a m -> 'b m

val (<*>) : ('a -> 'b) m -> 'a m -> 'b m

val (<*) : 'a m -> 'b m -> 'a m

val (*>) : 'a m -> 'b m -> 'b m

val catch : ('a -> 'a m) -> 'a m -> 'a m

val filter : ('a -> bool) -> 'a m -> 'a m

val of_list : 'a list -> 'a m

val sum : 'a list m -> 'a m

val msum : 'a m list -> 'a m

Boolean functions

val guard : bool -> unit m

val only_if : bool -> (unit -> 'a) -> 'a m

val optionally : ('a -> 'b m) -> 'a option -> 'b m

overwrites Applicative.optionally with a more general type. ATTENTION: changes semantics; None maps to zero (), not return ()!

val conditional : bool -> (unit -> 'a m) -> 'a m

overwrites Applicative.conditional with a more general type. ATTENTION: changes semantics; false maps to zero (), not return ()!

Miscellaneous

val transpose : 'a list m -> 'a m list

Generalises matrix transposition. This will loop infinitely if BasePlus.null cannot answer true for zeroes.