Module Option.Monad

include MonadPlus.T with type 'a m = 'a option

include BatInterfaces.Monad with type 'a m = 'a option

type 'a m = 'a option

val zero : unit -> 'a m

generates a neutral element with respect to plus

val plus : 'a m -> 'a m -> 'a m

add two monadic values

val null : 'a m -> bool

null x implies that x is zero. If you do not want to or cannot answer whether a given x is zero, then null x should raise an exception. I have provided this so that streams can be implemented more efficiently.

include Monad.S with type 'a m := 'a m

include BatInterfaces.Monad with type 'a m := 'a m

val bind : 'a m -> ( 'a -> 'b m ) -> 'b m

module Ap : Applicative.S with type 'a m := 'a m

include Applicative.S with type 'a m := 'a m

include Applicative.T with type 'a m := 'a m

val return : 'a -> 'a m

val (<*>) : ( 'a -> 'b ) m -> 'a m -> 'b m

val map : ( 'a -> 'b ) -> 'a m -> 'b m

map f x applies a pure function f to a monadic value x.

val fmap : ( 'a -> 'b ) m -> 'a -> 'b m

map f x applies a monadic function f to a pure value x.

val iter : ( 'a -> unit ) -> 'a m -> unit

iter f x applies a pure function f to a monadic value x and discards the result.

val fiter : ( 'a -> unit ) m -> 'a -> unit

iter f x applies a monadic function f to a pure value x and discards the result.

val (<@>) : ( 'a -> 'b ) -> 'a m -> 'b m

Alias for map. f <@> x applies a pure function f to a monadic value x.

val (<*>) : ( 'a -> 'b ) m -> 'a m -> 'b m

f <*> x applies a monadic function f to a monadic value x. This operator is usually used in conjunction with <@> in order to accumulate monadic parameters. Example: max <@> return 1 <*> return 2 will yield return 2. Here max has arity 2, so we use <*> to add a second monadic parameter.

val (<*) : 'a m -> 'b m -> 'a m

x <* y ignores the result of y.

val (*>) : 'a m -> 'b m -> 'b m

x <* y ignores the result of x.

val ignore : 'a m -> unit m

val map2 : ( 'a -> 'b -> 'c ) -> 'a m -> 'b m -> 'c m

map2 f x y is the same as the same as f <@> x <*> y.

val map3 : ( 'a -> 'b -> 'c -> 'd ) -> 'a m -> 'b m -> 'c m -> 'd m

map3 f x y z is the same as f <@> x <*> y <*> z.

val map4 : ( 'a -> 'b -> 'c -> 'd -> 'e ) -> 'a m -> 'b m -> 'c m -> 'd m -> 'e m

map4 f w x y z is the same as f <@> w <*> x <*> y <*> z.

List functions

val sequence : ( unit -> 'a m ) list -> 'a list m

val sequence_unit : ( unit -> unit m ) list -> unit m

val list_map : ( 'a -> 'b m ) -> 'a list -> 'b list m

like List.map but for monadic functions.

val list_iter : ( 'a -> unit m ) -> 'a list -> unit m

like List.iter but for monadic functions.

val list_filter : ( 'a -> bool m ) -> 'a list -> 'a list m

like List.filter but for monadic functions.

val list_filter_map : ( 'a -> 'b option m ) -> 'a list -> 'b list m

like List.filter_map but for monadic functions.

Option functions

val option_map : ( 'a -> 'b m ) -> 'a option -> 'b option m

like Option.map but for monadic functions.

Boolean function

val optional : bool -> ( unit -> 'a m ) -> 'a option m

optional true f executes f and applies Some to the result; optional false f returns None.

Tuple functions

val (>>=) : 'a m -> ( 'a -> 'b m ) -> 'b m

an operator for bind; The monadic bind operator passes the result of a monadic computation to the next monadic computation.

val (>=>) : ( 'a -> 'b m ) -> ( 'b -> 'c m ) -> 'a -> 'c m

a monadic function composition; f >=> g applies f to some argument and then g to the resulting value.

val (<=<) : ( 'b -> 'c m ) -> ( 'a -> 'b m ) -> 'a -> 'c m

a monadic function composition; g <=< f applies f to some argument and then g to the resulting value.

include module type of Ap.Infix

val (<@>) : ( 'a -> 'b ) -> 'a m -> 'b m

Alias for map. f <@> x applies a pure function f to a monadic value x.

val (<*>) : ( 'a -> 'b ) m -> 'a m -> 'b m

f <*> x applies a monadic function f to a monadic value x. This operator is usually used in conjunction with <@> in order to accumulate monadic parameters. Example: max <@> return 1 <*> return 2 will yield return 2. Here max has arity 2, so we use <*> to add a second monadic parameter.

val (<*) : 'a m -> 'b m -> 'a m

x <* y ignores the result of y.

val (*>) : 'a m -> 'b m -> 'b m

x <* y ignores the result of x.

module Bindings : sig ... end

include module type of Bindings

val let* : 'a m -> ( 'a -> 'b m ) -> 'b m

a binding operator for the monadic bind. let* y = x in f y corresponds to bind x f.

include module type of Ap.Bindings

val let+ : 'a m -> ( 'a -> 'b ) -> 'b m

binding operator for map

val and+ : 'a m -> 'b m -> ('a * 'b) m

binding operator for map

val join : 'a m m -> 'a m

val list_fold_left : ( 'a -> 'b -> 'a m ) -> 'a -> 'b list -> 'a m

like List.fold_left but for monadic functions

val list_fold_lefti : ( 'a -> int -> 'b -> 'a m ) -> 'a -> 'b list -> 'a m

like List.fold_lefti but for monadic functions

val list_fold_right : ( 'a -> 'b -> 'b m ) -> 'b -> 'a list -> 'b m

like List.fold_right but for monadic functions

val list_fold_righti : ( 'a -> int -> 'b -> 'b m ) -> 'b -> 'a list -> 'b m

like List.fold_righti but for monadic functions

module Tuple2 : sig ... end

module Tuple3 : sig ... end

module Tuple4 : sig ... end

module Tuple5 : sig ... end

module Infix : sig ... end

include module type of Infix

val (++) : 'a m -> 'a m -> 'a m

val (+?) : 'a option -> 'a m -> 'a m

val (>>=) : 'a m -> ( 'a -> 'b m ) -> 'b m

val (>=>) : ( 'a -> 'b m ) -> ( 'b -> 'c m ) -> 'a -> 'c m

val (<=<) : ( 'b -> 'c m ) -> ( 'a -> 'b m ) -> 'a -> 'c m

val (<@>) : ( 'a -> 'b ) -> 'a m -> 'b m

val (<*>) : ( 'a -> 'b ) m -> 'a m -> 'b m

val (<*) : 'a m -> 'b m -> 'a m

val (*>) : 'a m -> 'b m -> 'b m

val catch : ( 'a -> 'a m ) -> 'a m -> 'a m

val filter : ( 'a -> bool ) -> 'a m -> 'a m

Option functions

val of_option : 'a option -> 'a m

of_option (Some x) = return x; of_option None = zero ()

val optionally : ( 'a -> 'b m ) -> 'a option -> 'b m

optionally f (Some x) executes f x; optionally f None executes zero (). ATTENTION: this function overwrites Applicative.S.optionally generalising its type and changing its semantics! None maps to zero () instead of return ()!

val of_list : 'a list -> 'a m

val sum : 'a list m -> 'a m

val msum : 'a m list -> 'a m

Boolean functions

val guard : bool -> unit m

guard true does nothing; guard false yields zero ()). This can be used to render a computation conditional. Example: [let* () = guard (a = b) in f] Here [f] is only executed if [a = b].

val only_if : bool -> ( unit -> 'a ) -> 'a m

only_if true f = f (); only_if false f = zero ()

val conditional : bool -> ( unit -> 'a m ) -> 'a m

conditional true f executes f; conditional false f executes zero (). ATTENTION: this function overwrites Applicative.S.conditional generalising its type and changing its semantics! conditional false f maps to zero () instead of return ()!

Miscellaneous

val transpose : 'a list m -> 'a m list

Generalises matrix transposition. Relies on null.