Module State.CollectionState

Parameters

module T : sig ... end

module C : Collection.T

Signature

include Collection.T with type 'a m = T.s -> (T.s * 'a) C.m

include LazyPlus.T with type 'a m = T.s -> (T.s * 'a) C.m

include Monadlib.Monad.T with type 'a m = T.s -> (T.s * 'a) C.m

type 'a m = T.s -> (T.s * 'a) C.m

val zero : unit -> 'a m

val lplus : 'a m -> 'a m Stdlib.Lazy.t -> 'a m

val null : 'a m -> bool

null x implies that x is zero. If you do not want to or cannot answer whether a given x is zero, then null x should be false.

val difference : ('a -> 'a -> bool) -> 'a m -> 'a m -> 'a m

difference p xs ys removes all elements from xs which are less than or equal to some element in ys, according to the partial order p.

val unique : ?cmp:('a -> 'a -> bool) -> 'a m -> 'a m

unique eq xs removes duplicates according to the function cmp.

val maxima : ('a -> 'a -> bool) -> 'a m -> 'a m

maxima p xs leaves only the maximal elements according to the partial order p.

val nub : ('a -> 'a -> bool) -> 'a m -> 'a m

nub p xs ys is the same as maxima, but some values might be treated as greater than others depending on their position in the collection.

For instance, with lists, we treat elements that occur earlier in the list as always greater than elements that occur later, otherwise we use p. This is really just an optimisation: in order to retrieve any value from maxima p xs, the maxima function will have had to seen every value. Not so with nub.

include Monad.S with type 'a m := 'a m

include BatInterfaces.Monad with type 'a m := 'a m

val bind : 'a m -> ('a -> 'b m) -> 'b m

module Ap : Applicative.S with type 'a m := 'a m

include Applicative.S with type 'a m := 'a m

include Applicative.T with type 'a m := 'a m

val return : 'a -> 'a m

val (<*>) : ('a -> 'b) m -> 'a m -> 'b m

val map : ('a -> 'b) -> 'a m -> 'b m

map f x applies a pure function f to a monadic value x.

val fmap : ('a -> 'b) m -> 'a -> 'b m

map f x applies a monadic function f to a pure value x.

val iter : ('a -> unit) -> 'a m -> unit

iter f x applies a pure function f to a monadic value x and discards the result.

val (<@>) : ('a -> 'b) -> 'a m -> 'b m

Alias for map. f <@> x applies a pure function f to a monadic value x.

val (<*>) : ('a -> 'b) m -> 'a m -> 'b m

f <*> x applies a monadic function f to a monadic value x. This operator is usually used in conjunction with <@> in order to accumulate monadic parameters. Example: max <@> return 1 <*> return 2 will yield return 2. Here max has arity 2, so we use <*> to add a second monadic parameter.

val (<*) : 'a m -> 'b m -> 'a m

x <* y ignores the result of y.

val (*>) : 'a m -> 'b m -> 'b m

x <* y ignores the result of x.

val ignore : 'a m -> unit m

val map2 : ('a -> 'b -> 'c) -> 'a m -> 'b m -> 'c m

map2 f x y is the same as the same as f <@> x <*> y.

val map3 : ('a -> 'b -> 'c -> 'd) -> 'a m -> 'b m -> 'c m -> 'd m

map3 f x y z is the same as f <@> x <*> y <*> z.

val map4 : ('a -> 'b -> 'c -> 'd -> 'e) -> 'a m -> 'b m -> 'c m -> 'd m -> 'e m

map4 f w x y z is the same as f <@> w <*> x <*> y <*> z.

List functions

val sequence : (unit -> 'a m) list -> 'a list m

val sequence_unit : (unit -> unit m) list -> unit m

val list_map : ('a -> 'b m) -> 'a list -> 'b list m

like List.map but for monadic functions.

val list_iter : ('a -> unit m) -> 'a list -> unit m

like List.iter but for monadic functions.

val list_filter : ('a -> bool m) -> 'a list -> 'a list m

like List.filter but for monadic functions.

val list_filter_map : ('a -> 'b option m) -> 'a list -> 'b list m

like List.filter_map but for monadic functions.

Option functions

val option_map : ('a -> 'b m) -> 'a option -> 'b option m

like Option.map but for monadic functions.

val optionally : ('a -> unit m) -> 'a option -> unit m

optionally f (Some x) executes f x; optionally f None does nothing.

Boolean function

val optional : bool -> (unit -> 'a m) -> 'a option m

optional true f executes f and applies Some to the result; optional false f returns None.

val conditional : bool -> (unit -> unit m) -> unit m

conditional true f executes f; conditional false f does nothing.

Tuple functions

val split : ('a * 'b) m -> 'a m * 'b m

module Infix : sig ... end

include module type of Infix

val (>>=) : 'a m -> ('a -> 'b m) -> 'b m

an operator for bind; The monadic bind operator passes the result of a monadic computation to the next monadic computation.

val (>=>) : ('a -> 'b m) -> ('b -> 'c m) -> 'a -> 'c m

a monadic function composition; f >=> g applies f to some argument and then g to the resulting value.

val (<=<) : ('b -> 'c m) -> ('a -> 'b m) -> 'a -> 'c m

a monadic function composition; g <=< f applies f to some argument and then g to the resulting value.

include module type of Ap.Infix

val (<@>) : ('a -> 'b) -> 'a m -> 'b m

Alias for map. f <@> x applies a pure function f to a monadic value x.

val (<*>) : ('a -> 'b) m -> 'a m -> 'b m

f <*> x applies a monadic function f to a monadic value x. This operator is usually used in conjunction with <@> in order to accumulate monadic parameters. Example: max <@> return 1 <*> return 2 will yield return 2. Here max has arity 2, so we use <*> to add a second monadic parameter.

val (<*) : 'a m -> 'b m -> 'a m

x <* y ignores the result of y.

val (*>) : 'a m -> 'b m -> 'b m

x <* y ignores the result of x.

module Bindings : sig ... end

include module type of Bindings

val let* : 'a m -> ('a -> 'b m) -> 'b m

a binding operator for the monadic bind. let* y = x in f y corresponds to bind x f.

include module type of Ap.Bindings

val let+ : 'a m -> ('a -> 'b) -> 'b m

binding operator for map

val and+ : 'a m -> 'b m -> ('a * 'b) m

binding operator for map

val join : 'a m m -> 'a m

val list_fold_left : ('a -> 'b -> 'a m) -> 'a -> 'b list -> 'a m

like List.fold_left but for monadic functions

val list_fold_right : ('a -> 'b -> 'b m) -> 'b -> 'a list -> 'b m

like List.fold_right but for monadic functions

val read : T.s m

val write : T.s -> unit m

val cmp_on : ('a -> 'a -> bool) -> (T.s * 'a) -> (T.s * 'a) -> bool

cmp_on p is the usual product of the ordering on 'a with the ordering on T.s.

val lift : 'a C.m -> 'a m