Module Monadlib.Stream

module type S = sig ... end

Streams supporting fair and unbounded search.

module type StreamC = sig ... end

The union of streams and collections.

module Make (M : sig ... end) : sig ... end

Create a stream monad from an arbitrary inner monad, which computes the values discovered in each generation. This is pretty abstract, since we're not requiring that the inner monad is a collection. There is a constraint here, since we're strictly supposed to disallow a monad collection where order of elements matters. I think we can characterise this abstractly by saying that the plus operation on the inner monad must be commutative, but don't take my word for it!

module MakeStreamC (M : sig ... end) : sig ... end

Here, we create a stream monad from a definite collection monad Collection.T. The inner monad will be used to represent the generations in the stream. The order of elements in each generation should not matter, so you might want to use a set or a bag. If you want to live life on the edge, just remember that your code should not depend on the order of elements within generations (you can, of course, depend on the order that generations appear in the stream). You can enforce this constraint by performing, say, a sort on each generation.