Parameter MakeStreamC.1-M

include Applicative.T

type 'a m

val (<*>) : ( 'a -> 'b ) m -> 'a m -> 'b m

include Collection.T with type 'a m := 'a m

include LazyPlus.T with type 'a m := 'a m

include Monadlib.Monad.T with type 'a m := 'a m

val bind : 'a m -> ( 'a -> 'b m ) -> 'b m

val return : 'a -> 'a m

val zero : unit -> 'a m

val lplus : 'a m -> 'a m Stdlib.Lazy.t -> 'a m

val null : 'a m -> bool

null x implies that x is zero. If you do not want to or cannot answer whether a given x is zero, then null x should be false.

val difference : ( 'a -> 'a -> bool ) -> 'a m -> 'a m -> 'a m

difference p xs ys removes all elements from xs which are less than or equal to some element in ys, according to the partial order p.

val unique : ?cmp:( 'a -> 'a -> bool ) -> 'a m -> 'a m

unique eq xs removes duplicates according to the function cmp.

val maxima : ( 'a -> 'a -> bool ) -> 'a m -> 'a m

maxima p xs leaves only the maximal elements according to the partial order p.

val nub : ( 'a -> 'a -> bool ) -> 'a m -> 'a m

nub p xs ys is the same as maxima, but some values might be treated as greater than others depending on their position in the collection.

For instance, with lists, we treat elements that occur earlier in the list as always greater than elements that occur later, otherwise we use p. This is really just an optimisation: in order to retrieve any value from maxima p xs, the maxima function will have had to seen every value. Not so with nub.