Module 2-M.Infix

val (>>=) : 'a m -> ('a -> 'b m) -> 'b m

an operator for bind; The monadic bind operator passes the result of a monadic computation to the next monadic computation.

val (>=>) : ('a -> 'b m) -> ('b -> 'c m) -> 'a -> 'c m

a monadic function composition; f >=> g applies f to some argument and then g to the resulting value.

val (<=<) : ('b -> 'c m) -> ('a -> 'b m) -> 'a -> 'c m

a monadic function composition; g <=< f applies f to some argument and then g to the resulting value.

include module type of Ap.Infix

val (<@>) : ('a -> 'b) -> 'a m -> 'b m

Alias for map. f <@> x applies a pure function f to a monadic value x.

val (<*>) : ('a -> 'b) m -> 'a m -> 'b m

f <*> x applies a monadic function f to a monadic value x. This operator is usually used in conjunction with <@> in order to accumulate monadic parameters. Example: max <@> return 1 <*> return 2 will yield return 2. Here max has arity 2, so we use <*> to add a second monadic parameter.

val (<*) : 'a m -> 'b m -> 'a m

x <* y ignores the result of y.

val (*>) : 'a m -> 'b m -> 'b m

x <* y ignores the result of x.